+++ lib/bambora/v1/payment_resource.rb
@@ -78,6 +78,71 @@ module Bambora
 
       alias make_payment_with_payment_profile create_with_payment_profile
 
+      # Return payment.
+      #
+      # @example
+      #
+      #   client = Bambora::Rest::JSONClient(base_url: '...', api_key: '...', merchant_id: '...')
+      #   payments = Bambora::V1::PaymentResource(client: client)
+      #   payments.return_payment(
+      #     transaction_id: 1000341, amount: 50
+      #   )
+      #
+      # @param transaction_id [Integer] The transaction id
+      # @param amount [Float] A decimal value in dollars. Must be less than or equal to the original purchase amount.
+      # @param order_number [String] (Optional) A unique order number.
+      #
+      # @return [Hash] Indicating success or failure of the operation.
+      def return_payment(transaction_id:, amount:, order_number: nil)
+        data = { amount: amount, order_number: order_number }.compact
+        client.post(path: "#{sub_path}/#{transaction_id}/returns", body: data, api_key: api_key)
+      end
+
+      # Void a transaction. You can void payments, returns, pre-auths, and completions. It will
+      # cancel that transaction.
+      #
+      # @example
+      #
+      #   client = Bambora::Rest::JSONClient(base_url: '...', api_key: '...', merchant_id: '...')
+      #   payments = Bambora::V1::PaymentResource(client: client)
+      #   payments.void_payment(
+      #     transaction_id: 1000341, amount: 50
+      #   )
+      #
+      # @param transaction_id [Integer] The transaction id to void.
+      # @param amount [Float] A decimal value in dollars. Must be equal to the original purchase amount. You can void purchases as well as pre-auths and returns.
+      # @param order_number [String] (Optional) A unique order number.
+      #
+      # @return [Hash] Indicating success or failure of the operation.
+      def void(transaction_id:, amount:, order_number: nil)
+        data = { amount: amount, order_number: order_number }.compact
+        client.post(path: "#{sub_path}/#{transaction_id}/void", body: data, api_key: api_key)
+      end
+
+      alias void_payment void
+
+      # Complete a pre-authorized payment. The amount of the transaction to complete must be less
+      # than or equal to the original pre-auth amount.
+      #
+      # @example
+      #
+      #   client = Bambora::Rest::JSONClient(base_url: '...', api_key: '...', merchant_id: '...')
+      #   payments = Bambora::V1::PaymentResource(client: client)
+      #   payments.complete_preauth(
+      #     transaction_id: 1000341, amount: 50
+      #   )
+      #
+      # @param transaction_id [Integer] An integer identifier for the associated transaction.
+      # @param amount [Float] A decimal value in dollars. Uses up to two decimal places. Max value is account specific.
+      #   Default max value is 1000.
+      # @param order_number [String] (Optional) A unique order number.
+      #
+      # @return [Hash] Indicating success or failure of the operation.
+      def complete_preauth(transaction_id:, amount:, order_number: nil)
+        data = { amount: amount, order_number: order_number }.compact
+        client.post(path: "#{sub_path}/#{transaction_id}/completions", body: data, api_key: api_key)
+      end
+
       # Retrieve the details of a previously attempted payment.
       #
       #