+++ spec/bambora/v1/profile_resource_spec.rb
@@ -173,6 +173,150 @@ module Bambora
           ).to have_been_made.once
         end
       end
+
+      describe '#add_profile_card' do
+        let(:customer_code) { 'asdf1234' }
+        let(:data) do
+          {
+						token: {  
+							name:"John Doe",
+							code:"1eCe9480a7D94919997071a483505D17",
+						},
+          }
+        end
+
+        before do
+          stub_request(:post, "#{base_url}/v1/profiles/#{customer_code}/cards").with(
+            body: data.to_json.to_s,
+            headers: headers,
+          ).to_return(headers: response_headers, body: response_body.to_json.to_s)
+        end
+
+        it 'posts to the bambora api' do
+          subject.add_profile_card(customer_code: customer_code, data: data)
+
+          expect(
+            a_request(:post, "#{base_url}/v1/profiles/#{customer_code}/cards").with(
+              body: data.to_json.to_s,
+              headers: headers,
+            ),
+          ).to have_been_made.once
+        end
+      end
+
+      describe '#get' do
+        before do
+          stub_request(:get, "#{base_url}/v1/profiles/#{customer_code}")
+            .with(headers: headers)
+            .to_return(headers: response_headers, body: response_body.to_json.to_s)
+        end
+
+        let(:customer_code) { 'asdf1234' }
+        let(:headers) { { 'Authorization' => 'Passcode MTpmYWtla2V5' } }
+        let(:response_body) do
+          {
+            code: 1,
+            card: {},
+            billing: {},
+          }
+        end
+
+        it 'performs GET request for profile data' do
+          subject.get(customer_code: customer_code)
+
+          expect(
+            a_request(:get, "#{base_url}/v1/profiles/#{customer_code}")
+              .with(headers: headers),
+          ).to have_been_made.once
+        end
+      end
+
+      describe '#get_profile_cards' do
+        before do
+          stub_request(:get, "#{base_url}/v1/profiles/#{customer_code}/cards")
+            .with(headers: headers)
+            .to_return(headers: response_headers, body: response_body.to_json.to_s)
+        end
+
+        let(:customer_code) { 'asdf1234' }
+        let(:headers) { { 'Authorization' => 'Passcode MTpmYWtla2V5' } }
+        let(:response_body) do
+          {
+            code: 1,
+						message: "Operation Successful",
+						customer_code: 'aaa111',
+            card: [{
+              card_id: '1',
+              function: 'DEF',
+              name: 'Hup Podling',
+              number: '4030000010001234',
+              expiry_month: '12',
+              expiry_year: '23',
+              card_type: 'VI'
+            }],
+          }
+        end
+
+        it 'performs GET request for profile data' do
+          subject.get_profile_cards(customer_code: customer_code)
+
+          expect(
+            a_request(:get, "#{base_url}/v1/profiles/#{customer_code}/cards")
+              .with(headers: headers),
+          ).to have_been_made.once
+        end
+      end
+
+      describe '#update_profile_card' do
+        let(:customer_code) { 'asdf1234' }
+        let(:card_id) { 128 }
+
+        let(:data) do
+          {
+            card: {
+              expiry_month: '12',
+              expiry_year: '25',
+            },
+          }
+        end
+
+        before do
+          stub_request(:put, "#{base_url}/v1/profiles/#{customer_code}/cards/#{card_id}")
+            .with(headers: headers)
+            .to_return(headers: response_headers, body: response_body.to_json.to_s)
+        end
+
+        it 'posts to the bambora api' do
+          subject.update_profile_card(customer_code: customer_code, card_id: card_id, data: data)
+
+          expect(
+            a_request(:put, "#{base_url}/v1/profiles/#{customer_code}/cards/#{card_id}").with(
+              headers: headers,
+            ),
+          ).to have_been_made.once
+        end
+      end
+
+      describe '#delete_profile_card' do
+        let(:customer_code) { 'asdf1234' }
+        let(:card_id) { 128 }
+
+        before do
+          stub_request(:delete, "#{base_url}/v1/profiles/#{customer_code}/cards/#{card_id}")
+            .with(headers: headers)
+            .to_return(headers: response_headers, body: response_body.to_json.to_s)
+        end
+
+        it 'posts to the bambora api' do
+          subject.delete_profile_card(customer_code: customer_code, card_id: card_id)
+
+          expect(
+            a_request(:delete, "#{base_url}/v1/profiles/#{customer_code}/cards/#{card_id}").with(
+              headers: headers,
+            ),
+          ).to have_been_made.once
+        end
+      end
     end
   end
 end